Feat/broadcasting

+++ app/Http/Controllers/LikeController.php
@@ -2,10 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\NotificationEvent;
+use App\Events\PublicNotificationEvent;
 use App\Http\Requests\LikeCreateRequest;
-use App\Http\Requests\LikeDestroyRequest;
 use App\Http\Resources\LikeResource;
+use App\Http\Resources\NotificationResource;
 use App\Models\Like;
+use App\Models\Notification;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\JsonResponse;
 
 class LikeController extends Controller
@@ -16,12 +20,26 @@ class LikeController extends Controller
 			'quote_id'=> $request->quote_id,
 			'user_id' => $request->user_id,
 		]);
+		if ($request->author !== auth('sanctum')->user()->id) {
+			$notification = Notification::create([
+				'type'         => 'like',
+				'seen'         => false,
+				'user_id'      => auth('sanctum')->user()->id,
+				'post_author'  => $request->author,
+				'quote_id'     => $request->quote_id,
+			]);
+			event(new NotificationEvent(new NotificationResource($notification)));
+		}
+		event(new PublicNotificationEvent(new LikeResource($like), true));
 		return response()->json(['message'=>'success', 'like'=>new LikeResource($like)], 201);
 	}
 
-	public function destroy(LikeDestroyRequest $request, Like $like): JsonResponse
+	public function destroy(Like $like): JsonResponse
 	{
-		$like->delete();
-		return response()->json(['message'=>'like deleted'], 204);
+		if (Gate::allows('delete', $like)) {
+			$like->delete();
+			event(new PublicNotificationEvent(new LikeResource($like), false));
+			return response()->json(['message'=>'like deleted'], 204);
+		}
 	}
 }
