+++ app/Http/Controllers/MovieController.php
@@ -8,20 +8,15 @@ use App\Http\Resources\MovieResource;
 use App\Models\Genre;
 use App\Models\Movie;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Gate;
 
 class MovieController extends Controller
 {
-	public function index(Request $request): JsonResponse
+	public function index(): JsonResponse
 	{
-		$query = $request->query('query');
-		if ($query) {
-			$movies = Movie::filterByName($query)->with('quotes')->get();
-			return response()->json(['movies'=>MovieResource::collection($movies)]);
-		}
-		return response()->json(['message'=>'success', 'movies'=>MovieResource::collection(Movie::all())]);
+		$movies = Movie::where('user_id', auth()->user()->id)->orderBy('created_at', 'desc')->get();
+		return response()->json(['movies'=>MovieResource::collection($movies)]);
 	}
 
 	public function show(Movie $movie): JsonResponse
@@ -49,33 +44,37 @@ class MovieController extends Controller
 
 	public function update(MovieUpdateRequest $request, Movie $movie): JsonResponse
 	{
-		if ($request->hasFile('image')) {
-			$request->validate(['image'=>'image|mimes:png,jpg']);
-			$imagePath = public_path('storage/images/' . $movie->image);
-			if (File::exists($imagePath)) {
-				File::delete($imagePath);
+		if (Gate::allows('update', $movie)) {
+			if ($request->hasFile('image')) {
+				$request->validate(['image'=>'image|mimes:png,jpg']);
+				$imagePath = public_path('storage/images/' . $movie->image);
+				if (File::exists($imagePath)) {
+					File::delete($imagePath);
+				}
+				$imagePath = $request->file('image')->store('public/images');
+				$movie->update(
+					[
+						'image'      => basename($imagePath),
+						...$request->validated(),
+					]
+				);
+			} else {
+				$movie->update($request->validated());
 			}
-			$imagePath = $request->file('image')->store('public/images');
-			$movie->update(
-				[
-					'image'      => basename($imagePath),
-					...$request->validated(),
-				]
-			);
-		} else {
-			$movie->update($request->validated());
+			$genreData = json_decode($request->genre, true);
+			$genreIds = array_column($genreData, 'id');
+			$genres = Genre::whereIn('id', $genreIds)->get();
+			$movie->genres()->detach();
+			$movie->genres()->attach($genres);
+			return response()->json(['message'=>'movie updated', 'movie'=>new MovieResource($movie)], 201);
 		}
-		$genreData = json_decode($request->genre, true);
-		$genreIds = array_column($genreData, 'id');
-		$genres = Genre::whereIn('id', $genreIds)->get();
-		$movie->genres()->detach();
-		$movie->genres()->attach($genres);
-		return response()->json(['message'=>'movie updated', 'movie'=>new MovieResource($movie)], 201);
 	}
 
 	public function destroy(Movie $movie): JsonResponse
 	{
-		$movie->delete();
-		return response()->json(['message'=>'movie deleted', 'movie'=>$movie], 202);
+		if (Gate::allows('delete', $movie)) {
+			$movie->delete();
+			return response()->json(['message'=>'movie deleted', 'movie'=>$movie], 202);
+		}
 	}
 }
